lesson23/requirements.txt
aiohttp==3.9.1
openai==1.6.1
python-dotenv==1.0.0
asyncio 